requirements.txt
azure-identity>=1.12.0
azure-mgmt-securityinsight>=2.0.0
azure-mgmt-resource>=23.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tabulate>=0.9.0
